Remove old auto-import code
