+++ src/main/kotlin/org/rust/ide/utils/import/ImportCandidatesCollector2.kt
@@ -57,10 +57,10 @@ import org.rust.stdext.mapToSet
  * - Std
  * - Workspace
  * - Dependencies
- * See [ImportCandidate2.compareTo].
+ * See [ImportCandidate.compareTo].
  */
 object ImportCandidatesCollector2 {
-    fun getImportCandidates(context: ImportContext2, targetName: String): List<ImportCandidate2> {
+    fun getImportCandidates(context: ImportContext2, targetName: String): List<ImportCandidate> {
         val itemsPaths = context.getAllModPaths()
             .flatMap { context.getAllItemPathsInMod(it, targetName) }
         return context.convertToCandidates(itemsPaths)
@@ -70,7 +70,7 @@ object ImportCandidatesCollector2 {
         context: ImportContext2,
         prefixMatcher: PrefixMatcher,
         processedElements: MultiMap<String, RsElement>
-    ): List<ImportCandidate2> {
+    ): List<ImportCandidate> {
         val modPaths = context.getAllModPaths()
         val allNames = modPaths.flatMapTo(hashSetOf()) { it.mod.visibleItems.keys }
         val nameToPriority = prefixMatcher.sortMatching(allNames)
@@ -91,11 +91,11 @@ object ImportCandidatesCollector2 {
      * * given [resolvedMethods] don't refer to any trait
      * * if at least one trait related to [resolvedMethods] is already in scope
      */
-    fun getImportCandidates(scope: RsElement, resolvedMethods: List<MethodResolveVariant>): List<ImportCandidate2>? =
+    fun getImportCandidates(scope: RsElement, resolvedMethods: List<MethodResolveVariant>): List<ImportCandidate>? =
         getTraitImportCandidates(scope, resolvedMethods.map { it.source })
 
-    fun getTraitImportCandidates(scope: RsElement, sources: List<TraitImplSource>): List<ImportCandidate2>? {
-        val traits = ImportCandidatesCollector.collectTraitsToImport(scope, sources)
+    fun getTraitImportCandidates(scope: RsElement, sources: List<TraitImplSource>): List<ImportCandidate>? {
+        val traits = collectTraitsToImport(scope, sources)
             ?: return null
         val traitsPaths = traits.mapNotNullToSet { it.asModPath() }
 
@@ -105,7 +105,15 @@ object ImportCandidatesCollector2 {
         return context.convertToCandidates(itemsPaths)
     }
 
-    private fun getImportCandidates(context: ImportContext2, target: RsQualifiedNamedElement): List<ImportCandidate2> {
+    private fun collectTraitsToImport(scope: RsElement, sources: List<TraitImplSource>): List<RsTraitItem>? {
+        val traits = sources.mapNotNull { source ->
+            if (source.isInherent) return null
+            source.requiredTraitInScope
+        }
+        return if (traits.filterInScope(scope).isNotEmpty()) null else traits
+    }
+
+    private fun getImportCandidates(context: ImportContext2, target: RsQualifiedNamedElement): List<ImportCandidate> {
         val name = if (target is RsFile) {
             target.modName
         } else {
@@ -115,11 +123,11 @@ object ImportCandidatesCollector2 {
             .filter { it.qualifiedNamedItem.item == target }
     }
 
-    fun findImportCandidate(context: ImportContext2, target: RsQualifiedNamedElement): ImportCandidate2? =
+    fun findImportCandidate(context: ImportContext2, target: RsQualifiedNamedElement): ImportCandidate? =
         getImportCandidates(context, target).firstOrNull()
 }
 
-private fun ImportContext2.convertToCandidates(itemsPaths: List<ItemUsePath>): List<ImportCandidate2> =
+private fun ImportContext2.convertToCandidates(itemsPaths: List<ItemUsePath>): List<ImportCandidate> =
     itemsPaths
         .groupBy { it.toItemWithNamespace() }
         .mapValues { (item, paths) -> filterForSingleItem(paths, item) }
@@ -140,7 +148,7 @@ private fun ImportContext2.convertToCandidates(itemsPaths: List<ItemUsePath>): L
                     val qualifiedItem = QualifiedNamedItem2(itemPsi, path.path, path.crate)
                     val importInfo = qualifiedItem.toImportInfo(rootDefMap, rootModData, path.needExternCrate)
                     val isRootPathResolved = isRootPathResolved(importInfo.usePath)
-                    ImportCandidate2(qualifiedItem, importInfo, isRootPathResolved)
+                    ImportCandidate(qualifiedItem, importInfo, isRootPathResolved)
                 }
             }
         }
